a kind of quick image denoising method
LMS algorithm is used for image noise reduction, and the method is improved, the effect of the improved method is tested, and the noise reduction effect obtained based on the DnCNN network method is compared.     And the effect of the improved method is close to the DnCNN method, and if you are interested in  DnCNN, please refer to "Remove Noise from Color Image Using Pretrained Neural Network" in matlab.   The advantage of my method is that I don't need any high-level GPU to pretrain the net and Deep Learning Toolbox to support  application, you just need a essential CPU.